pyproject.toml
[tool.poetry]
name = "coffee-records"
version = "0.1.0"
description = "Espresso shot tracking application"
authors = ["Scott"]
packages = [{ include = "coffee_records" }]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0"
sqlalchemy = "^2.0"
pydantic = "^2.0"
pydantic-settings = "^2.0"
psycopg2-binary = "^2.9"
pyyaml = "^6.0"
gunicorn = "^22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
black = "^24.0"
mypy = "^1.0"
types-pyyaml = "^6.0"
sqlalchemy-stubs = "^0.4"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["test"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
